Remove FieldAnalysisEmit
Have Emit inherit directly from Design
