+++ pyaedt/emit.py
@@ -3,14 +3,16 @@ from __future__ import absolute_import
 import warnings
 
 from pyaedt import generate_unique_project_name
-from pyaedt.application.AnalysisEmit import FieldAnalysisEmit
+from pyaedt.application.Design import Design
+from pyaedt.modeler.schematic import ModelerEmit
+from pyaedt.emit_core.Couplings import CouplingsEmit
 from pyaedt.emit_core import EMIT_MODULE
 from pyaedt.emit_core import EmitConstants
 from pyaedt.emit_core.results.results import Results
 from pyaedt.generic.general_methods import pyaedt_function_handler
 
 
-class Emit(FieldAnalysisEmit, object):
+class Emit(Design, object):
     """Provides the Emit application interface.
 
     Parameters
@@ -130,13 +132,12 @@ class Emit(FieldAnalysisEmit, object):
         }
         """Default Emit units."""
 
-        FieldAnalysisEmit.__init__(
+        Design.__init__(
             self,
             "EMIT",
             projectname,
             designname,
             solution_type,
-            setup_name,
             specified_version,
             non_graphical,
             new_desktop_session,
@@ -146,6 +147,8 @@ class Emit(FieldAnalysisEmit, object):
             port=port,
             aedt_process_id=aedt_process_id,
         )
+        self._modeler = ModelerEmit(self)
+        self._couplings = CouplingsEmit(self)
 
         if self._aedt_version >= "2023.1":
             self._emit_api = EMIT_MODULE.EmitApi()
@@ -155,6 +158,32 @@ class Emit(FieldAnalysisEmit, object):
             """''Result'' object for the selected design."""
 
             self.__emit_api_enabled = True
+        
+        # Update modeler IDs now, after units are initialized.
+        #self.modeler.components.refresh_all_ids()
+
+
+    @property
+    def modeler(self):
+        """Modeler.
+
+        Returns
+        -------
+        pyaedt.modeler.schematic.ModelerEmit
+            Design oModeler
+        """
+        return self._modeler
+
+    @property
+    def couplings(self):
+        """Emit Couplings.
+
+        Returns
+        -------
+        pyaedt.emit_core.Couplings.CouplingsEmit
+            Couplings within the EMIT Design
+        """
+        return self._couplings
 
     @pyaedt_function_handler()
     def __enter__(self):
